# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Amazon Bid Engine API
  description: |
    Amazon広告自動入札提案エンジンのREST API。

    ## 認証
    すべてのエンドポイント（/health を除く）では `X-API-Key` ヘッダーによる認証が必要です。

    ## Retool統合
    このAPIはRetool Cloudからの利用を想定しています。
    - Base URL: Cloud Runのデプロイ先URL
    - Authentication: Header Auth (`X-API-Key`)
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://your-service.a.run.app
    description: Production (Cloud Run)
  - url: http://localhost:8080
    description: Local Development
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for authentication
...
paths:
...
  /admin/executions:
    get:
      summary: 実行履歴一覧
      description: 入札エンジンの実行履歴を取得
      tags:
        - Executions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [RUNNING, SUCCESS, FAILED]
        - name: mode
          in: query
          schema:
            type: string
            enum: [SHADOW, APPLY]
      responses:
        '200':
          description: 成功
  /admin/executions/{executionId}/asin-summary:
    get:
      summary: 実行のASIN別サマリー
      tags:
        - Executions
      parameters:
        - name: executionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
  /admin/executions/{executionId}/keyword-details:
    get:
      summary: 実行のキーワード詳細
      tags:
        - Executions
      parameters:
        - name: executionId
          in: path
          required: true
          schema:
            type: string
        - name: asin
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 成功
...